Chat reducer combined with word and picture
