+++ src/components/ChatApp.js
@@ -4,12 +4,13 @@ import { sendMessage } from "../store/chat/chat";
 import "../index.scss";
 class Chat extends Component {
   render() {
+    console.log("PROPS", this.props);
     const { feed, sendMessage } = this.props;
     return (
       <div class="chat">
         <h1>Blah Blah Chat</h1>
         <ul>
-          {feed.map((entry) => (
+          {feed.chat.map((entry) => (
             <li>{entry.text}</li>
           ))}
         </ul>
