+++ src/index.js
@@ -6,8 +6,7 @@ import Chat from "./components/ChatApp";
 import * as serviceWorker from "./serviceWorker";
 import { BrowserRouter as Router } from "react-router-dom";
 import { Provider } from "react-redux";
-//import store from "./store";
-import { store } from "./store/chat/chat";
+import { store } from "./store";
 import "milligram";
 
 ReactDOM.render(
