+++ src/store/index.js
@@ -2,11 +2,12 @@ import { createStore, applyMiddleware, compose } from "redux";
 
 import ReduxThunk from "redux-thunk";
 import reducer from "./rootReducer";
+import { messageMiddleware } from "./chat/chat";
 
 const composeEnhancers = window.__REDUX_DEVTOOLS_EXTENSION_COMPOSE__ || compose;
 
-const enhancer = composeEnhancers(applyMiddleware(ReduxThunk));
+const enhancer = composeEnhancers(
+  applyMiddleware(ReduxThunk, messageMiddleware)
+);
 
-const store = createStore(reducer, enhancer);
-
-export default store;
+export const store = createStore(reducer, enhancer);
