+++ src/store/rootReducer.js
@@ -1,10 +1,10 @@
 import { combineReducers } from "redux";
 import word from "./word/reducer";
 import picture from "./picture/reducer";
-import messageReducer from "./chat/chat";
+import chat from "./chat/reducer";
 
 export default combineReducers({
   word,
   picture,
-  messageReducer,
+  chat,
 });
